requirements.txt
python-osc>=1.8.1
pylive @ git+https://github.com/ideoforms/pylive.git
python-rtmidi>=1.5.0
numpy>=1.24.0
openai>=1.3.0  # For natural language processing
python-dotenv>=1.0.0  # For environment variable management 